% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radians.R
\name{get_radians}
\alias{get_radians}
\title{get_radians}
\usage{
get_radians(degrees)
}
\arguments{
\item{degrees}{angle value in degrees to be transformed}
}
\value{
A double describing radian angle.
}
\description{
Convert angle in degrees to radians.
}
\examples{
get_radians(degrees=90)
}
