% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trajectory.R
\name{get_trajectory}
\alias{get_trajectory}
\title{get_trajectory}
\usage{
get_trajectory(o_x, o_y, d_x, d_y, od_pair)
}
\arguments{
\item{o_x}{Location in x of origin.}

\item{o_y}{Location in y of origin.}

\item{d_x}{Location in x of destination.}

\item{d_y}{Location in y of destination.}

\item{od_pair}{Character string naming OD-pair.}
}
\value{
A tibble of coordinate pairs representing the, origin, destination
and control point of the asymmetric curve.
}
\description{
This function takes cartesian coordinates defining origin and
destination locations returns a tibble representing a path for an asymmetric
bezier curve, which  curves towards the destination location. The tibble has
three rows representing an origin, destination and control point for the
bezier curve. The parameterisation follows that published in
\href{https://openaccess.city.ac.uk/id/eprint/538/}{Wood et al. 2011}.
}
\examples{
outward <- get_trajectory(o_x=0,o_y=0,d_x=1,d_y=0, "out")
inward <- get_trajectory(o_x=1,o_y=0,d_x=0,d_y=0, "in")
}
