% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivers.R
\docType{data}
\name{rivers}
\alias{rivers}
\title{River outline collected from OSM}
\format{
An sf data.frame with 1 observation and 2 variables:
\describe{
\item{id}{num row identifier}
\item{geometry}{MULTIPOLYGON representing river outline}
}
}
\source{
\url{https://www.openstreetmap.org/}
}
\usage{
rivers
}
\description{
River outline of Thames collected from OSM and represented in sf geometry, snapped to bounds of London bikeshare.
}
\keyword{datasets}
