% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trajectory.R
\name{get_radians}
\alias{get_radians}
\title{Create path defining asymmetric bezier curve. Parameterisation follows
that published in Wood et al. 2011. doi: 10.3138/carto.46.4.239.}
\usage{
get_radians(degrees)
}
\arguments{
\item{data}{A df defining an OSGB OD pair (e.g. cartesian coordinates)
and weight.}
}
\value{
A df of coord pairs representing asymmetric curve.
}
\description{
Create path defining asymmetric bezier curve. Parameterisation follows
that published in Wood et al. 2011. doi: 10.3138/carto.46.4.239.
}
